pyproject.toml
[tool.poetry]
name = "python-ab2-hw-kafka-ilystsov"
version = "0.1.0"
description = ""
authors = ["ilya <lystsov.ilya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
confluent-kafka = "^2.3.0"
fastavro = "^1.9.4"
requests = "^2.31.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.4.2"
pytest = "^8.2.0"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
